Change the strided copy to switch on data size not data type. Move to header so we can reduce on the enabled types. Setup type reduction for Concat now that it's using this implementation.
